refactor: 적절한 도메인이 있는 곳으로 클래스를 이동
